finish making transaction module

+++ src/transaction/transaction.service.ts
@@ -1,6 +1,7 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { PrismaService } from 'src/common/prisma/prisma';
 import { AddTransactionInput } from './dto/add-transaction';
+import { UpdateTransactionInput } from './dto/edit-transaction';
 
 
 @Injectable()
@@ -16,8 +17,12 @@ export class TransactionService {
             if (!portfolio) {
                 throw new NotFoundException('Portfolio not found');
             }
+
+            if (portfolio.authorId !== userId) {
+                throw new UnauthorizedException('No access!!');
+            }
             
-            const transaction = await this.prisma.transaction.create({
+            return this.prisma.transaction.create({
                 data: {
                     coin: dto.coin,
                     operation: dto.operation,
@@ -29,16 +34,37 @@ export class TransactionService {
                 }
             });
 
-            return transaction;
-
         } catch (error) {
             throw error;
         }
     }
 
-    public async getAllTransaction() {
+    public async editTransaction(dto: UpdateTransactionInput, userId: number) {
         try {
+            const transaction = await this.getOneTransaction(dto.id)
             
+            if (!transaction) {
+                throw new NotFoundException('Portfolio not found');
+            }
+
+            if (transaction.authorId !== userId) {
+                throw new UnauthorizedException('No access!!');
+            }
+            
+            return this.prisma.transaction.update({
+                where: {
+                    id: dto.id
+                },
+                data: {
+                    coin: dto.coin,
+                    operation: dto.operation,
+                    price: dto.price,
+                    spent: dto.spent,
+                    date: dto.date,
+                    authorId: +userId,
+                }
+            });
+
         } catch (error) {
             throw error;
         }
@@ -61,17 +87,13 @@ export class TransactionService {
         }
     }
 
-    public async editTransaction() {
-        try {
-            
-        } catch (error) {
-            throw error;
-        }
-    }
-
-    public async deleteTransaction(id: number) {
+    public async deleteTransaction(id: number, userId: number) {
         try {
             const transactionToDelete = await this.getOneTransaction(id);
+
+            if (transactionToDelete.authorId !== userId) {
+                throw new UnauthorizedException('No access!!');
+            }
 
             await this.prisma.transaction.delete({
                 where: {
